move share extension view from Views to App
