+++ apple/OmnivoreKit/Sources/Views/LocalText.swift
@@ -10,12 +10,12 @@ public enum LocalText {
   static let registrationViewSignInHeadline = localText(key: "registrationViewSignInHeadline")
   static let registrationViewSignUpHeadline = localText(key: "registrationViewSignUpHeadline")
   public static let registrationViewHeadline = localText(key: "registrationViewHeadline")
-  static let networkError = localText(key: "error.network")
-  static let genericError = localText(key: "error.generic")
+  public static let networkError = localText(key: "error.network")
+  public static let genericError = localText(key: "error.generic")
   static let invalidCredsLoginError = localText(key: "loginError.invalidCreds")
-  static let saveArticleSavedState = localText(key: "saveArticleSavedState")
-  static let saveArticleProcessingState = localText(key: "saveArticleProcessingState")
-  static let extensionAppUnauthorized = localText(key: "extensionAppUnauthorized")
+  public static let saveArticleSavedState = localText(key: "saveArticleSavedState")
+  public static let saveArticleProcessingState = localText(key: "saveArticleProcessingState")
+  public static let extensionAppUnauthorized = localText(key: "extensionAppUnauthorized")
   static let dismissButton = localText(key: "dismissButton")
   static let usernameValidationErrorInvalid = localText(key: "username.validation.error.invalidPattern")
   static let usernameValidationErrorTooShort = localText(key: "username.validation.error.tooshort")
